Add comment above IncidentDetails.Alerts field explaining behaviors
For the Webhook object sent from PagerDuty, the "alerts" key is not a JSON array
of fully hydrated objects. Instead, only the "alert_key" JSON key is set within
each object, which might be unclear based on the type we use in the Go client.

This change adds a comment above the Alerts field to indicate it's not a slice
of fully hydrated `IncidentAlert` values, and that only `AlertKey` is set.